% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusion_graph.R
\name{add_fp_edge}
\alias{add_fp_edge}
\title{Add Fusion-Preventing Edge}
\usage{
add_fp_edge(from, to, graph, ...)
}
\arguments{
\item{from}{the node the edge starts at}

\item{to}{the node the edge ends at}

\item{graph}{a fusion graph}

\item{...}{additional arguments to addEdge}
}
\description{
Add a fusion-preventing edge to a fusion graph.
}

