% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusion_graph.R
\name{add_fusion_edges}
\alias{add_fusion_edges}
\title{Add Edges To The Fusion Graph}
\usage{
add_fusion_edges(graph, i, nodes)
}
\arguments{
\item{graph}{a fusion graph}

\item{i}{index of the current block}

\item{nodes}{a list of dependency information for the nodes}
}
\description{
Add Edges To The Fusion Graph
}

