% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{as_code_block}
\alias{as_code_block}
\title{Convert To Code Block}
\usage{
as_code_block(x, f)
}
\arguments{
\item{x}{a list of code objects}

\item{f}{a function to apply to each code block before combining}
}
\description{
This function converts a list of code objects to a single code block,
preserving the order of the list.
}

