% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusion_graph.R
\name{blockify}
\alias{blockify}
\title{Group a list of code objects into blocks, while keeping the specified
classes as independent blocks.}
\usage{
blockify(x, as_block)
}
\arguments{
\item{x}{a list of code objects or a code block}

\item{as_block}{classes which should be treated as independent blocks}
}
\description{
Group a list of code objects into blocks, while keeping the specified
classes as independent blocks.
}

