% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependence_collector.R
\name{collect_deps}
\alias{collect_deps}
\title{Collect Dependence Information}
\usage{
collect_deps(x)
}
\arguments{
\item{x}{a block of code}
}
\description{
This function collects the dependence information for a block of code.
}

