% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop_fusion.R
\name{combine_for_loops}
\alias{combine_for_loops}
\title{Combine a List of for-loops}
\usage{
combine_for_loops(loops)
}
\arguments{
\item{loops}{an ordered list of for-loops}
}
\description{
This function combines a list of for-loops in the same order they appear in
the list, using the header of the first loop.
}

