% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop_fusion.R
\name{combine_fusion_graph_nodes}
\alias{combine_fusion_graph_nodes}
\title{Combine Nodes in a Fusion Graph}
\usage{
combine_fusion_graph_nodes(g, partition, node = paste0(partition, collapse =
  ""))
}
\arguments{
\item{g}{a fusion graph}

\item{partition}{a vector of nodes to combine}

\item{node}{a name for the combined node}
}
\description{
This function combines the specified nodes in a fusion graph, preserving any
fusion-preventing edges.
}

