% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_helpers.R
\name{edge_attr}
\alias{edge_attr}
\title{Extract and Simplify an Edge Attribute}
\usage{
edge_attr(graph, attr)
}
\arguments{
\item{graph}{a graph}

\item{attr}{the name of a edge attribute}
}
\description{
Extract the specified edge attribute for all edges, and simplify the
resulting list to a vector whenever possible.
}

