% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop_fusion.R
\name{fuse_component_graph}
\alias{fuse_component_graph}
\title{Compute and Fuse a Component Graph}
\usage{
fuse_component_graph(graph, nodes, component)
}
\arguments{
\item{graph}{a fusion graph}

\item{nodes}{a list of code blocks}

\item{component}{a component name ("parallel" or "serial")}
}
\description{
This function computes the requested component graph and fuses it, then
updates the fusion graph and list of code blocks appropriately.
}

