% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusion_graph.R
\name{fusion_graph}
\alias{fusion_graph}
\title{Generate a Fusion Graph}
\usage{
fusion_graph(x)
}
\arguments{
\item{x}{a function or code block}
}
\description{
This function generates the fusion graph for the specified block of code.
}

