% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop_fusion.R
\name{greedy_partition}
\alias{greedy_partition}
\title{Greedy Partitioning Algorithm}
\usage{
greedy_partition(gc)
}
\arguments{
\item{gc}{component graph}
}
\description{
This function partitions a component graph so that each partition is a set
of fusable nodes.
}
\details{
The original algorithm is described in (Allen 1986), but was modified here
to suit the loop fusion strategy described in (Kennedy 1993).
}

