% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusion_graph.R
\name{header_equal}
\alias{header_equal}
\title{Check whether two for-loops have equal headers.}
\usage{
header_equal(x, y)
}
\arguments{
\item{x}{the first for-loop}

\item{y}{the second for-loop}
}
\description{
Check whether two for-loops have equal headers.
}

