% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{in_which}
\alias{in_which}
\title{Which Elements Contain The Value?}
\usage{
in_which(x, l)
}
\arguments{
\item{x}{a value}

\item{l}{a list}
}
\description{
This function returns the names of all elements that contain the specified
value.
}

