% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop_fusion.R
\name{loop_fusion}
\alias{loop_fusion}
\title{Loop Fusion}
\usage{
loop_fusion(x)
}
\arguments{
\item{x}{a function or code block}
}
\description{
This function applies the loop fusion algorithm described in (Kennedy 1993)
to the specified function or code block.
}

