% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop_fusion.R
\name{necessary_edges}
\alias{necessary_edges}
\title{Necessary Edges Algorithm}
\usage{
necessary_edges(g, gc, type)
}
\arguments{
\item{g}{original graph}

\item{gc}{component graph}

\item{type}{component type}
}
\description{
Necessary Edges Algorithm
}

