% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusion_graph.R
\name{new_fusion_graph}
\alias{new_fusion_graph}
\title{Create a New Fusion Graph}
\usage{
new_fusion_graph()
}
\description{
Create an empty directed graph with default node and edge data suitable for
use in loop fusion.
}

