% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_helpers.R
\name{node_attr}
\alias{node_attr}
\title{Extract and Simplify a Node Attribute}
\usage{
node_attr(graph, attr)
}
\arguments{
\item{graph}{a graph}

\item{attr}{the name of a node attribute}
}
\description{
Extract the specified node attribute for all nodes, and simplify the
resulting list to a vector whenever possible.
}

