% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_helpers.R
\name{subgraph}
\alias{subgraph}
\title{Get a Subgraph}
\usage{
subgraph(graph, nodes)
}
\arguments{
\item{graph}{a graph}

\item{nodes}{the nodes to keep}
}
\description{
Get the subgraph of a graph. Behavior is similar to subGraph from the graph
package, but the data defaults for the graph are preserved.
}

