% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency_collector.R
\name{collect_deps}
\alias{collect_deps}
\title{Collect variable dependency information for a code block.}
\usage{
collect_deps(x)
}
\description{
Collect variable dependency information for a code block.
}

